Updated: update.sh

+++ i113_panda_update.sh
@@ -7,7 +7,11 @@ echo $service_image_name
 
 docker --version
 
+if docker ps | grep -- $service_name > /dev/null
+then
 docker stop $service_name
 docker rm $service_name
 docker rmi $service_image_name
-
+else
+echo no such container
+fi
